+++ src/Renderers/RangeSliderRenderer.cpp
@@ -32,6 +32,9 @@ namespace tgui
 {
     TGUI_RENDERER_PROPERTY_COLOR(RangeSliderRenderer, SelectedTrackColor, {})
     TGUI_RENDERER_PROPERTY_COLOR(RangeSliderRenderer, SelectedTrackColorHover, {})
+
+    TGUI_RENDERER_PROPERTY_TEXTURE(RangeSliderRenderer, TextureSelectedTrack)
+    TGUI_RENDERER_PROPERTY_TEXTURE(RangeSliderRenderer, TextureSelectedTrackHover)
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////